Moving interfaces that produce no code to a single file
