--- src/Measure.ts
@@ -1,3 +1,0 @@
-export interface IMeasure {
-  chords: any[];
-}
