--- src/Part.ts
@@ -1,4 +1,0 @@
-export interface IPart {
-  chord?: string;
-  lyric?: string;
-}
