--- src/Section.ts
@@ -1,6 +1,0 @@
-import { Line } from "./Line";
-
-export interface ISection {
-  lines: Line[];
-  name: string;
-}
